refactors folder structure and adds share function
