+++ web/index.html
@@ -313,6 +313,7 @@
 
     <script src="https://cdnjs.cloudflare.com/ajax/libs/ace/1.4.2/ace.js"></script>
     <script src="dist/nice-select2.js"></script>
+    <script type="module" src="assets/js/theme.js"></script>
     <script type="module" src="assets/js/services/modes.js"></script>
     <script type="module" src="assets/js/services/examples.js"></script>
     <script type="module" src="assets/js/utils/render-functions.js"></script>
